更新

+++ soinsu5.cu
@@ -24,14 +24,14 @@ __host__ int GCD(int a, int b)
 	}
 }
 
-__global__ void kernel(long long *A, int *d_B)
+__global__ void kernel(int *A, int *d_B)
 {
-	long long i = blockIdx.x * blockDim.x + threadIdx.x;
-	long long j = blockIdx.y * blockDim.y + threadIdx.y;
+	int i = blockIdx.x * blockDim.x + threadIdx.x;
+	int j = blockIdx.y * blockDim.y + threadIdx.y;
 	int a = i - j;
 	if(i >= __powf(*A,0.5) + 1 && j >= __powf(*A,0.5) + 1 && a > 1 && i < *A && j < *A){
 		if(i^2 % *A == j^2 % *A){
-			d_B[sizeof(d_B) / sizeof(long long)] = a;
+			d_B[sizeof(d_B) / sizeof(int)] = a;
 		}
 	}
 }
@@ -39,13 +39,13 @@ __global__ void kernel(long long *A, int *d_B)
 int main(){
 	clock_t t1, t2;
 	t1 = times_clock();
-    long long *d_target, A = target;
+    int *d_target, A = target;
 	int *d_B;
 	int B[SIZE];
 	int i, j, k;
-    cudaMalloc((void**)&d_target,sizeof(long long));
+    cudaMalloc((void**)&d_target,sizeof(int));
 	cudaMalloc((void**)&d_B,sizeof(int)*SIZE);
-	cudaMemcpy(d_target,&A,sizeof(long long),cudaMemcpyHostToDevice);
+	cudaMemcpy(d_target,&A,sizeof(int)),cudaMemcpyHostToDevice);
 	cudaMemcpy(d_B,&B,sizeof(int)*SIZE,cudaMemcpyHostToDevice);
 	dim3 block(32,32);
 	dim3 grid((A+31)/32,(A+31)/32);
